Share settlement copy, tier SFX, and entry option pills.
Delete Stage-local duplicates of CrashTicketSettlement status strings, RoundTheater tier sounds, and CrashRoundEntry OptionGroup.
