+++ src/components/round-theater/round-theater.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { useReplayClock } from "@/hooks/use-replay-clock";
 import {
   useRoundTheater,
@@ -10,13 +10,9 @@ import {
   type TheaterView,
 } from "@/hooks/use-round-theater";
 import { useTheaterPlayerTicket } from "@/hooks/use-theater-player-ticket";
+import { useTheaterTierSounds } from "@/hooks/use-theater-tier-sounds";
 import { formatDeskDollarsAmount } from "@/lib/desk-dollars";
-import { getTierCloseProgress } from "@/lib/round-replay";
-import {
-  ENTRY_LEVERAGE_TIERS_BPS,
-  formatLeverageBps,
-  type CrashTicket,
-} from "@/lib/margin-call-crash";
+import { formatLeverageBps, type CrashTicket } from "@/lib/margin-call-crash";
 import {
   formatNextRoundHandoff,
   formatTimelineCountdownLabel,
@@ -382,7 +378,7 @@ function ReplayStage({
     restartNonce,
   });
 
-  useTierSoundEffects({
+  useTheaterTierSounds({
     crashPointBps: hero.crashPointBps,
     progress: clock.progress,
     isComplete: clock.isComplete,
@@ -508,58 +504,4 @@ function ExpiredStage({
       <TicketTape entries={live.tape?.entries ?? []} />
     </div>
   );
-}
-
-function useTierSoundEffects(options: {
-  crashPointBps: bigint;
-  progress: number;
-  isComplete: boolean;
-  enabled: boolean;
-  restartNonce: number;
-  playerTierBps: bigint | null;
-}) {
-  // The per-frame work is two number comparisons against precomputed,
-  // ascending close thresholds; the log math runs once per Crash Point.
-  const closeThresholds = useMemo(
-    () =>
-      ENTRY_LEVERAGE_TIERS_BPS.flatMap((tier) => {
-        const closeAt = getTierCloseProgress(tier, options.crashPointBps);
-        if (closeAt === null) return [];
-        return [
-          {
-            closeAt,
-            isPlayerTier:
-              options.playerTierBps !== null && tier === options.playerTierBps,
-          },
-        ];
-      }).sort((a, b) => a.closeAt - b.closeAt),
-    [options.crashPointBps, options.playerTierBps]
-  );
-  const firedCountRef = useRef(0);
-  const crashedRef = useRef(false);
-
-  useEffect(() => {
-    firedCountRef.current = 0;
-    crashedRef.current = false;
-  }, [options.restartNonce, options.crashPointBps]);
-
-  useEffect(() => {
-    if (!options.enabled) return;
-    while (
-      firedCountRef.current < closeThresholds.length &&
-      closeThresholds[firedCountRef.current].closeAt <= options.progress
-    ) {
-      const threshold = closeThresholds[firedCountRef.current];
-      firedCountRef.current += 1;
-      // The player's own close rings the register instead of the desk chime.
-      if (threshold.isPlayerTier) getTheaterAudio().playWinRegister();
-      else getTheaterAudio().playTierClose();
-    }
-    if (options.isComplete && !crashedRef.current) {
-      crashedRef.current = true;
-      const audio = getTheaterAudio();
-      audio.playCrashBell();
-      audio.playPhoneRing();
-    }
-  }, [closeThresholds, options.enabled, options.isComplete, options.progress]);
 }
